cloud: always wrap cloud storage implementations
This is a minor change to always wrap external storage implementations
with a wrapper that constructs readers and writers with additional
implementation-agnostic implementations.

The point of this minor change is to make it easier to add additional
wrapping readers and writers that aren't related to the existing
limiters.

Release note: None