streamingccl: check heartbeat expiration from re-loaded job record
Previously, because of the unused function argument, we were checking
the heartbeat expiration on the copy of the job we had that was
initially passed to us from resume. This copy is never updated. As a
result, the job would always time out.

Now, we check the correct job object and the job does not time out.

Release note: None