backupccl: unit test for restore debugging
The big_entry.json is a dump of the restore spans created by the tpcc
incremental roachtest.  Some of these restore spans are _extremely_
slow to process.

I recommend either running this test in a gce worker for fast access
to gcs or to modify the restore span entry to use local files.

Release note: None