storage: fix data race in MVCCPredicateDeleteRange
This fixes two bugs in MVCCPredicateDeleteRange that would result in
incorrect logical operations being published to rangefeeds.

When publishing a logical operation, the caller indicates whether the
buffers provide for the Key and EndKey of the operation are safe for
concurrent use by setting a `Safe` boolean field in the published
operation.

MVCCPredicateDeleteRange calls either MVCCDeleteRangeUsingTombstone or
mvccPutInternal, both of which publish logical opts with Safe set to
true.

Unfortunatelely, the code in MVCCPredicateDeleteRange _did_ invalidate
the keys passed to these functions.

As a result, it was possible for rangefeeds to observe:

1. Range keys with incorrect bounds,
2. Missed pointed deletions events,
3. Unnecessarily duplicated point deletions, and
4. Point deletions that never occurred (but were part of range deletions).

Additionally, since we expect to be able to read the values for a
logical operation back out of storage, rangefeeds may also encounter
errors as they failed to read back the value for an erroneous events.

Fixes #112733

Epic: none

Release note(bug fix): Fix a bug that would result in incorrect
physical replication results for users of the private preview physical
replication feature.