streamingccl: use correct bounds for range key SSTs
Previously, we were erroneously calling Next() on the end key of our
SST even though the end key already represented an exclusive bounds.

This Next() meant that the bounds of our SSTs were slightly
incorrect. In most cases, this was harmless, but in the case of a
range boundary that aligned directly with the _actual_ bounds of the
SST, it could result in us attempting to split an SST into 2 with the
RHS ending up empty.

Here, we fix that and also make our SST splitting a bit more robust to
bad input. It now errors if the LHS or RHS ends up empty despite a
split key inside the given bounds. It also handles split keys outside
of the bounds. Neither of these things should happen unless our
contract with other parts of the system changes.

Fixes #112846

Release note (bug fix): Fix a bug that could prevent phyiscal
replication from advancing in the face of some range deletion
operations.