skip: cleanup skip output
Previously, becuase of how our lists of interfaces had been passed
around, we logged skips with square brackets around them in some cases
because we were actually passing a single []interface{} rather than
expanding the var args.

Here, I make it a little easier to avoid that mistake.

I also normalize the behaviour of skip.IgnoreLint with
skip.IgnoreLintf such that they both ignore the env var. I think this
is OK since this is typically used when the test is skipped because
some other pre-requisite is not met.

Release note: None