backupccl: pessimistic fix for online restore SST bounds
Currently, we have to assume that the EndKey of a backup file manifest
entry is inclusive. The restore span generation code assumed this but
the online restore code did not. Here, we account for that by
producing spans that end at EndKey.Next().

This is a bit pessimistic since most of the time the end key of the
file manifest entry is in fact exclusive. There are many cases where
we know this for sure when constructing the manifest entry, but we
don't currently include that information in the manifest itself so by
the time of restore, we can no longer be sure.

Release note: None