rangefeed: defer unregistration until stream is removed from manager
Unregistering the registration from the processor has the side-effect of
closing the underlying memory budget. We don't want that to happen until
all of the events have been cleared from the queue.

Here, we pass the unregistration callback up via the Disconnector and
then the stream manager calls it when removing the stream.

Epic: none
Release note: None